Changed organization + added classes for maps
I finished the classes to create the map. 
All we have to do is a driver for the map to show as an example.

Also, I changed the organization of the files because if the files are separated from one another, then the compiler will ask for the full directory if you want to include a file. That's not good because directories vary from one computer to another.

I'm currently working on the Map Loader.